feat: 대시보드 기능 추가 및 리팩토링
- 인프라 섹션(InfraSection) 추가: 노드 및 리소스 상태를 시각적으로 표시하는 컴포넌트 구현.
- 모니터링 섹션(MonitoringSection) 추가: 서비스의 메트릭을 기반으로 한 상위 N개 항목을 표시하는 컴포넌트 구현.
- 나의 서비스 카드(MyServiceCards) 추가: 사용자의 서비스를 카드 형태로 표시하고 드래그 가능하도록 구현.
- 사용자 테이블(UserTable) 추가: 회원 목록을 테이블 형태로 표시하는 컴포넌트 구현.
- 대시보드 페이지(page.tsx) 리팩토링: 기존 코드 정리 및 새로운 컴포넌트 통합.
- 대시보드 관련 API 훅(useGetDashboardSummary, useGetDashboardTopUsers) 수정: enabled 파라미터 추가로 조건부 호출 가능하도록 개선.
- CSS 스타일 추가 및 수정: 새로운 컴포넌트에 필요한 스타일 정의 및 기존 스타일 정리.
