+++ package.json
@@ -22,6 +22,7 @@
     "@tanstack/react-query": "^5.100.5",
     "@xyflow/react": "^12.10.2",
     "clsx": "^2.1.1",
+    "framer-motion": "^12.40.0",
     "ky": "^1.14.3",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
